+++ lib/zip/central_directory.rb
@@ -26,10 +26,13 @@ module Zip
       @entry_set.each { |entry| entry.write_c_dir_entry(io) }
       eocd_offset = io.tell
       cdir_size = eocd_offset - cdir_offset
-      has_zip64_entry = @entry_set.any? { |entry| entry.extra['Zip64'] }
-      if has_zip64_entry || cdir_offset > 0xFFFFFFFF || cdir_size > 0xFFFFFFFF || @entry_set.size > 0xFFFF
-        write_64_e_o_c_d(io, cdir_offset, cdir_size)
-        write_64_eocd_locator(io, eocd_offset)
+      if ::Zip.write_zip64_support
+        need_zip64_eocd = cdir_offset > 0xFFFFFFFF || cdir_size > 0xFFFFFFFF || @entry_set.size > 0xFFFF
+        need_zip64_eocd ||= @entry_set.any? { |entry| entry.extra['Zip64'] }
+        if need_zip64_eocd
+          write_64_e_o_c_d(io, cdir_offset, cdir_size)
+          write_64_eocd_locator(io, eocd_offset)
+        end
       end
       write_e_o_c_d(io, cdir_offset, cdir_size)
     end
@@ -43,7 +46,7 @@ module Zip
         @entry_set ? [@entry_set.size, 0xFFFF].min : 0,
         [cdir_size, 0xFFFFFFFF].min,
         [offset, 0xFFFFFFFF].min,
-        @comment ? @comment.length : 0
+        @comment ? @comment.bytesize : 0
       ]
       io << tmp.pack('VvvvvVVv')
       io << @comment
@@ -93,7 +96,7 @@ module Zip
       @size_in_bytes                                = Entry.read_zip_64_long(buf)
       @cdir_offset                                  = Entry.read_zip_64_long(buf)
       @zip_64_extensible                            = buf.slice!(0, buf.bytesize)
-      raise ZipError, "Zip consistency problem while reading eocd structure" unless buf.size == 0
+      raise Error, "Zip consistency problem while reading eocd structure" unless buf.size == 0
     end
 
     def read_e_o_c_d(buf) #:nodoc:
@@ -105,19 +108,19 @@ module Zip
       @size_in_bytes                                = Entry.read_zip_long(buf)
       @cdir_offset                                  = Entry.read_zip_long(buf)
       comment_length                                = Entry.read_zip_short(buf)
-      @comment                                      = if comment_length <= 0
+      @comment                                      = if comment_length.to_i <= 0
                                                         buf.slice!(0, buf.size)
                                                       else
                                                         buf.read(comment_length)
                                                       end
-      raise ZipError, "Zip consistency problem while reading eocd structure" unless buf.size == 0
+      raise Error, "Zip consistency problem while reading eocd structure" unless buf.size == 0
     end
 
     def read_central_directory_entries(io) #:nodoc:
       begin
         io.seek(@cdir_offset, IO::SEEK_SET)
       rescue Errno::EINVAL
-        raise ZipError, "Zip consistency problem while reading central directory entry"
+        raise Error, "Zip consistency problem while reading central directory entry"
       end
       @entry_set = EntrySet.new
       @size.times do
@@ -137,7 +140,7 @@ module Zip
 
     def get_e_o_c_d(buf) #:nodoc:
       sig_index = buf.rindex([END_OF_CDS].pack('V'))
-      raise ZipError, "Zip end of central directory signature not found" unless sig_index
+      raise Error, "Zip end of central directory signature not found" unless sig_index
       buf = buf.slice!((sig_index + 4)..(buf.bytesize))
 
       def buf.read(count)
@@ -162,9 +165,9 @@ module Zip
 
     def get_64_e_o_c_d(buf) #:nodoc:
       zip_64_start = buf.rindex([ZIP64_END_OF_CDS].pack('V'))
-      raise ZipError, "Zip64 end of central directory signature not found" unless zip_64_start
+      raise Error, "Zip64 end of central directory signature not found" unless zip_64_start
       zip_64_locator = buf.rindex([ZIP64_EOCD_LOCATOR].pack('V'))
-      raise ZipError, "Zip64 end of central directory signature locator not found" unless zip_64_locator
+      raise Error, "Zip64 end of central directory signature locator not found" unless zip_64_locator
       buf = buf.slice!((zip_64_start + 4)..zip_64_locator)
 
       def buf.read(count)
@@ -179,7 +182,7 @@ module Zip
       @entry_set.each(&proc)
     end
 
-    # Returns the number of entries in the central directory (and 
+    # Returns the number of entries in the central directory (and
     # consequently in the zip archive).
     def size
       @entry_set.size
@@ -189,7 +192,7 @@ module Zip
       cdir = new
       cdir.read_from_stream(io)
       return cdir
-    rescue ZipError
+    rescue Error
       return nil
     end