+++ lib/zip/errors.rb
@@ -1,8 +1,16 @@
 module Zip
-  class ZipError < StandardError; end
-  class ZipEntryExistsError < ZipError; end
-  class ZipDestinationFileExistsError < ZipError; end
-  class ZipCompressionMethodError < ZipError; end
-  class ZipEntryNameError < ZipError; end
-  class ZipInternalError < ZipError; end
+  class Error < StandardError; end
+  class EntryExistsError < Error; end
+  class DestinationFileExistsError < Error; end
+  class CompressionMethodError < Error; end
+  class EntryNameError < Error; end
+  class InternalError < Error; end
+
+  # Backwards compatibility with v1 (delete in v2)
+  ZipError = Error
+  ZipEntryExistsError = EntryExistsError
+  ZipDestinationFileExistsError = DestinationFileExistsError
+  ZipCompressionMethodError = CompressionMethodError
+  ZipEntryNameError = EntryNameError
+  ZipInternalError = InternalError
 end