+++ lib/zip/output_stream.rb
@@ -24,14 +24,17 @@ module Zip
 
     # Opens the indicated zip file. If a file with that name already
     # exists it will be overwritten.
-    def initialize(fileName, stream=false)
+    def initialize(file_name, stream=false)
       super()
-      @fileName = fileName
-      if stream
-        @output_stream = ::StringIO.new
-      else
-        @output_stream = ::File.new(@fileName, "wb")
-      end
+      @file_name = file_name
+      @output_stream = if stream
+                         iostream = @file_name.dup
+                         iostream.reopen
+                         iostream.rewind
+                         iostream
+                       else
+                         ::File.new(@file_name, "wb")
+                       end
       @entry_set = ::Zip::EntrySet.new
       @compressor = ::Zip::NullCompressor.instance
       @closed = false
@@ -43,19 +46,19 @@ module Zip
     # stream is passed to the block and closed when the block
     # returns.
     class << self
-      def open(fileName)
-        return new(fileName) unless block_given?
-        zos = new(fileName)
+      def open(file_name)
+        return new(file_name) unless block_given?
+        zos = new(file_name)
         yield zos
       ensure
         zos.close if zos
       end
 
-	    # Same as #open but writes to a filestream instead
-      def write_buffer
-        zos = new('', true)
+      # Same as #open but writes to a filestream instead
+      def write_buffer(io = ::StringIO.new(''))
+        zos = new(io, true)
         yield zos
-        return zos.close_buffer
+        zos.close_buffer
       end
     end
 
@@ -79,35 +82,36 @@ module Zip
       @output_stream
     end
 
-	  # Closes the current entry and opens a new for writing.
+    # Closes the current entry and opens a new for writing.
     # +entry+ can be a ZipEntry object or a string.
-    def put_next_entry(entryname, comment = nil, extra = nil, compression_method = Entry::DEFLATED,  level = Zlib::DEFAULT_COMPRESSION)
-      raise ZipError, "zip stream is closed" if @closed
-      if entryname.kind_of?(Entry)
-        new_entry = entryname
+    def put_next_entry(entry_name, comment = nil, extra = nil, compression_method = Entry::DEFLATED, level = Zip.default_compression)
+      raise Error, "zip stream is closed" if @closed
+      if entry_name.kind_of?(Entry)
+        new_entry = entry_name
       else
-        new_entry = Entry.new(@fileName, entryname.to_s)
+        new_entry = Entry.new(@file_name, entry_name.to_s)
       end
-      new_entry.comment = comment if !comment.nil?
-      if (!extra.nil?)
+      new_entry.comment = comment unless comment.nil?
+      unless extra.nil?
         new_entry.extra = ExtraField === extra ? extra : ExtraField.new(extra.to_s)
       end
-      new_entry.compression_method = compression_method if !compression_method.nil?
+      new_entry.compression_method = compression_method unless compression_method.nil?
       init_next_entry(new_entry, level)
       @current_entry = new_entry
     end
 
     def copy_raw_entry(entry)
       entry = entry.dup
-      raise ZipError, "zip stream is closed" if @closed
-      raise ZipError, "entry is not a ZipEntry" if !entry.kind_of?(Entry)
+      raise Error, "zip stream is closed" if @closed
+      raise Error, "entry is not a ZipEntry" unless entry.is_a?(Entry)
       finalize_current_entry
       @entry_set << entry
-      src_pos = entry.local_entry_offset
+      src_pos = entry.local_header_offset
       entry.write_local_entry(@output_stream)
       @compressor = NullCompressor.instance
       entry.get_raw_input_stream do |is|
         is.seek(src_pos, IO::SEEK_SET)
+        ::Zip::Entry.read_local_entry(is)
         IOExtras.copy_stream_n(@output_stream, is, entry.compressed_size)
       end
       @compressor = NullCompressor.instance
@@ -123,10 +127,10 @@ module Zip
       @current_entry.size = @compressor.size
       @current_entry.crc = @compressor.crc
       @current_entry = nil
-      @compressor = NullCompressor.instance
+      @compressor = ::Zip::NullCompressor.instance
     end
 
-    def init_next_entry(entry, level = Zlib::DEFAULT_COMPRESSION)
+    def init_next_entry(entry, level = Zip.default_compression)
       finalize_current_entry
       @entry_set << entry
       entry.write_local_entry(@output_stream)
@@ -135,10 +139,13 @@ module Zip
 
     def get_compressor(entry, level)
       case entry.compression_method
-        when Entry::DEFLATED then Deflater.new(@output_stream, level)
-        when Entry::STORED   then PassThruCompressor.new(@output_stream)
-      else raise ZipCompressionMethodError,
-        "Invalid compression method: '#{entry.compression_method}'"
+      when Entry::DEFLATED then
+        ::Zip::Deflater.new(@output_stream, level)
+      when Entry::STORED then
+        ::Zip::PassThruCompressor.new(@output_stream)
+      else
+        raise ::Zip::CompressionMethodError,
+              "Invalid compression method: '#{entry.compression_method}'"
       end
     end
 
@@ -163,10 +170,13 @@ module Zip
     end
 
     public
+
     # Modeled after IO.<<
     def << (data)
       @compressor << data
+      self
     end
+
   end
 end